% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inst.R
\name{add_inst}
\alias{add_inst}
\title{Add an instrument}
\usage{
add_inst(filepath, f0 = 440, fs = 44100, inst_lab, save_dir)
}
\arguments{
\item{filepath}{string: the location of a brief sound file that the synthesizer will try to emulate.}

\item{f0}{numeric: the fundamental frequency of the input sound in Hertz. This defaults to 440 (A above middle C in concert tuning), so the user should either provide a sound file that matches this pitch or determine the appropriate \code{f0} of the sound file they are using.}

\item{fs}{integer: the sampling frequency (or frame rate) of the sound file in Hertz. This defaults to 44100, but 48000 is not uncommon. The user should check the frame rate of their sound file beforehand.}

\item{inst_lab}{string: the label that will be used to refer to the synthesizer parameters generated. Note that "bass", "piano", "string", and "wind" are already assigned by default and cannot be overwritten. This string is used to index a \code{list}, so verify that \code{inst_lab} is a valid list index.}

\item{save_dir}{string: the location where R will save the custom instrument sounds as \code{instruments_custom.rda}. If unspecified, R will save in the current working directory.}
}
\value{
-1 if the user tries to assign an \code{inst_lab} that already exists. Otherwise, returns nothing.
}
\description{
Takes a user-specified sound (stored at \code{filepath} on the local machine), infers synthesizer parameters from the sound, and saves the synthesizer parameters to be used later (e.g., via \code{notes()}) under the label \code{inst_lab}.
}
\examples{
\dontshow{wplayback <- FALSE}
# Generate a 250 Hz sawtooth wave and save it in the local folder
# using the save.wave() function from library(audio)
fs <- 44100
f0 <- 250
dur <- 2
t <- seq(0, dur, length.out = dur * fs)
sawtooth <- 2 * ((250 * t) \%\% 1) - 1
audio::save.wave(what = sawtooth, where = here::here("example.wav"))

# Open the newly saved sound file and use it to specify 
# a new synthesizer sound under the label "saw".
add_inst(here::here("example.wav"), f0 = f0, fs = fs, inst_lab = "saw")

# Use the newly generated sound in a short melody
melody <- c(40, 38, 36, 38, 40, 40, 40)
wplay(notes(melody, dur = 0.25, vol = 0.3, inst_lab = "saw"), normalize = FALSE)

}
