% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bach.R
\name{bach}
\alias{bach}
\title{Bach Prelude 2}
\usage{
bach(
  spd = 6,
  fs = 44100,
  inst_lab = "piano",
  decay = FALSE,
  decay_rate = -1,
  want_df = FALSE,
  stereo = FALSE,
  progbar = TRUE
)
}
\arguments{
\item{spd}{numeric: speed of playback (default: \code{6})}

\item{fs}{integer: sampling rate (default: \code{44100}.  Sampling rate is dependent on user audio system, typically 44100 or 48000).}

\item{inst_lab}{string: instrument used (default:\code{"piano"}).}

\item{decay}{logical: if \code{TRUE}, volume of individual notes decays over duration of note (default: \code{FALSE})}

\item{decay_rate}{numeric: The decay rate of the sound. Larger -> Faster decay. (negative numbers result in sound growing louder over duration of note)}

\item{want_df}{logical: if \code{TRUE} returns data frame in long format (default: \code{FALSE})}

\item{stereo}{logical: if \code{TRUE} returns stereo output, separating notes by hand (default = \code{FALSE})}

\item{progbar}{logical: if \code{TRUE} progress bar enabled}
}
\value{
vector or matrix
}
\description{
Generates playable vector of the first half of Prelude 2 (BWV 871) from The Well-Tempered Clavier, Book II
}
\examples{
\dontrun{
wplay_controls <- TRUE
wplay(bach(decay = TRUE))
}
}
