% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_count.R
\name{create_count}
\alias{create_count}
\title{4 count snare}
\usage{
create_count(dur, fs = 44100)
}
\arguments{
\item{dur}{numeric: duration of the output in seconds}

\item{fs}{integer: sampling rate of the output in Hertz}
}
\value{
numeric vector representing the waveform of the 4 count sound
}
\description{
function to create 4 count of the specified dur
}
