% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{Decompose a waveform into sine waves}
\usage{
decomp(y, n_harm = 12, fs = 44100, f0 = 440)
}
\arguments{
\item{y}{n_samples length numeric vector}

\item{n_harm}{integer:  number of harmonics to analyze (i.e., the number of filters to use)}

\item{fs}{integer: The sampling rate of \code{y}}

\item{f0}{numeric: The center frequency of the lowest band-pass filter}
}
\value{
Decomposition matrix. Each row represents one component sine wave in \code{y}, assuming \code{y} is composed of harmonically related sines. Summing the columns will closely approximate the original input \code{y}.
}
\description{
Decomposes the mono waveform \code{y} using a bank of harmonically related band-pass filters (i.e., they have center frequencies that are integer multiples of \code{f0}). Each row of the output is a filtered version of the input. If the input is a single complex tone with pitch \code{f0}, this function isolates the harmonics in the complex.
}
\examples{
\dontrun{
# Create a signal that is a sum of harmonically related sinusoids
fs <- 44100
dur <- 2
t <- seq(0, dur, length.out = dur * fs)
f0 <- 200
n_harm <- 8
y <- sin(2 * pi * f0 * t)
for (k in 2:n_harm) {
  y <- y + sin(2 * pi * k * f0 * t)
}

# Separate the sinusoids
yd <- decomp(y, n_harm = n_harm, fs = fs, f0 = f0)

# First row is the sinusoid at the fundamental frequency
plot(yd[1, 1:2000], type = "l")
}
}
