% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tag_audio.R
\name{html_tag_audio}
\alias{html_tag_audio}
\title{Create audio tag (helper function / Internal Use)}
\usage{
html_tag_audio(file_path, type = "wav")
}
\arguments{
\item{file_path}{The path of the audio file}

\item{type}{The file type (wav/mp3)}
}
\value{
html audio tag
}
\description{
Used in \code{wplay()} when knitting an .Rmd file to html.  Inserts audio tags / creates audio player within document.
}
\examples{
\dontshow{wplayback <- FALSE}
html_tag_audio("foo")

}
