% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scale.R
\name{make_scale}
\alias{make_scale}
\alias{major_semitones}
\alias{minor_semitones}
\alias{major_scale}
\alias{minor_scale}
\title{Make a scale}
\usage{
make_scale(tonic_pkey, n_octaves = 1, scale_type = "major")

major_semitones()

minor_semitones()

major_scale(tonic_pkey, n_octaves = 1)

minor_scale(tonic_pkey, n_octaves = 1)
}
\arguments{
\item{tonic_pkey}{numeric: the root pitch from which to begin the scale (if \code{tonic_pkey = 0} generates semitones for major/minor scale of specified octaves)}

\item{n_octaves}{integer: the number of octaves ascending from \code{tonic_pkey}}

\item{scale_type}{string: "major" (default) or "minor"; specify the type of scale to create}
}
\value{
an integer vector of piano key numbers corresponding to the created scale
}
\description{
Generates piano key designations for (ascending) scale at specified tonic and number of octaves
}
\section{Functions}{
\itemize{
\item \code{major_semitones}: generates semitones from tonic for each note in 1 octave of an ascending major scale

\item \code{minor_semitones}: generates semitones from tonic for each note in 1 octave of an ascending minor scale

\item \code{major_scale}: generates an ascending major scale from the tonic

\item \code{minor_scale}: generates an ascending minor scale from the tonic
}}

\examples{
\dontshow{wplayback <- FALSE}
x <- major_scale(49)
wplay(notes(x, dur = 1 / 8))

}
