% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkey_to_hz.R
\name{pkey_to_hz}
\alias{pkey_to_hz}
\alias{hz_to_sec}
\title{piano key number to frequency (hz) to soundwave}
\usage{
pkey_to_hz(pkey)

hz_to_sec(hz, dur = 1, fs = 44100)
}
\arguments{
\item{pkey}{numeric: pitch expressed in piano key numbers (ideal piano).  40 corresponds to middle C / C4.}

\item{hz}{numeric: frequency of soundwave}

\item{dur}{integer: duration of the sound in seconds.}

\item{fs}{integer: sampling rate (default: \code{44100}.  Sampling rate is dependent on user audio system, typically 44100 or 48000).}
}
\value{
\itemize{
\item \code{pkey_to_hz()}: numeric frequency .
\item \code{hz_to_sec()}: numeric vector (playable with \code{audio::play()}, \code{cplay()}, or \code{wplay()})
}
}
\description{
\code{pkey_to_hz()} and \code{hz_to_sec()} are usually used in combination.
\itemize{
\item \code{pkey_to_hz()} maps piano key number to hz (i.e. input = pitch, output = frequency) using "idealized standard piano" frequencies.
\item \code{hz_to_sec()} generates a playable soundwave vector at the specified frequency and duration.
}
}
\section{Functions}{
\itemize{
\item \code{hz_to_sec}: generates playable vector with specified frequency (\code{hz})
}}

\examples{
\dontshow{wplayback <- FALSE}
# 49 is the piano key number for A440 (i.e. should return 440hz).
pkey_to_hz(49)

# vector for A440
A440 <- hz_to_sec(440)
wplay(A440)

# play middle C / C4
wplay(hz_to_sec(pkey_to_hz(40)))

}
