% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repitch.R
\name{repitch}
\alias{repitch}
\title{Repitch your audio}
\usage{
repitch(y, shift)
}
\arguments{
\item{y}{numeric vector representing the waveform to be repitched}

\item{shift}{numeric: the number of semitones to pitch shift \code{y} by. A value of 0 will return the input \code{y} unchanged. Negative values will result in slower, lower pitched outputs; positive values will result in faster, higher pitched outputs. This argument need not be an integer.}
}
\value{
a numeric vector that represents the pitch-shifted version of \code{y}
}
\description{
Takes a single-channel (mono) waveform \code{y} and shifts it by \code{shift} semitones.
\code{shift} may be positive or negative, or non-integer. Note that shifting pitch is accomplished via time dilation so the new note will be shorter for positive values of shift and longer for negative values.
}
\examples{
\dontshow{wplayback <- FALSE}
# Generate pure tone at 420 Hz
t <- seq(0, 1, length.out = 44100)
y <- sin(2 * pi * 420 * t)

# Play it out loud
wplay(y)

# Play it again after pitch-shifting down two semitones
wplay(repitch(y, -2))

}
