% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resynth.R
\name{resynth}
\alias{resynth}
\title{Synthesize amplitude-modulated complex tones}
\usage{
resynth(hz, dur, env, amp, phi, fs_old = 44100, fs_new = 44100)
}
\arguments{
\item{hz}{numeric: the frequency of the sound to be synthesized}

\item{dur}{numeric: the duration of the output sound in seconds}

\item{env}{numeric vector representing the envelope (amplitude over time) of the output sound at equidistant points spanning its entire duration.}

\item{amp}{numeric vector representing the amplitude of each harmonic in the output, starting from the fundamental frequency. If it is shorter than \code{phi}, then the missing entries will default to 1.}

\item{phi}{numeric vector representing the phase of each harmonic in the output, starting from the fundamental frequency. If it is shorter than \code{amp}, then the missing entries will default to 0.}

\item{fs_old}{integer: sampling frequency of the original sound used to infer \code{env}, \code{amp}, and \code{phi}. \strong{Now obsolete given the current synthesizer design}.}

\item{fs_new}{integer: sampling frequency of the output sound.}
}
\value{
n_samples length numeric vector representing the sound given by the above parameters.
}
\description{
Synthesize tones using parameters to specify envelope, relative amplitude of harmonics, and relative phase of harmonics. Generally, these parameters will be derived from pre-recorded sounds, and this function will be called by \code{snd_inst()}.
}
\examples{
\dontshow{wplayback <- FALSE}
# Specify the parameters by hand for a 360 Hz complex tone with linear decay
w <- resynth(hz = 360, dur = 1, env = c(1, 0), 
             amp = rev(seq(0, 1, length.out = 6)), 
             phi = rep(0, 6))

# Play it out loud
wplay(w)

}
