% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_num.R
\name{scale_num}
\alias{scale_num}
\title{Map your data to a scale}
\usage{
scale_num(
  x,
  tonic_pkey = 40,
  n_octaves = 2,
  scale_type = "major",
  new_scale = NULL,
  by_quantile = FALSE,
  under_hood = FALSE
)
}
\arguments{
\item{x}{numeric vector: variable to be binned and mapped to scale}

\item{tonic_pkey}{numeric: the root pitch from which to begin the scale (if \code{tonic_pkey = 0} generates semitones for major/minor scale of specified octaves)}

\item{n_octaves}{integer: the number of octaves ascending from \code{tonic_pkey}}

\item{scale_type}{string: "major" (default) or "minor"; specify the type of scale to create}

\item{new_scale}{numeric vector: vector of piano key number designations (can be fractional) to map bins to (if specified, overrides \code{tonic_pkey} and \code{scaletype} arguments).}

\item{by_quantile}{logical: if \code{TRUE}, bins values using equally-spaced quantiles as bin breakpoints (i.e. roughly equal number of \code{x}-values fall in each bin)}

\item{under_hood}{logical: if \code{TRUE}, returns list of most objects used to construct the mapping.}
}
\value{
either a numerical vector of piano key numbers (if \code{under_hood = FALSE}; default), or a list of objects used to construct the mapping (if \code{under_hood = TRUE}).
}
\description{
Place \code{x} values in equally-spaced bins (i.e. like a histogram's bins), then maps bins to specified scale (default scale is C4 major, 2 octaves).  Number of bins = number of notes in specified scale (i.e. dependent on \code{n_octaves}).
}
\examples{
\dontshow{wplayback <- FALSE}
scale_num(mtcars$mpg)

# looking under the hood:
x <- scale_num(mtcars$mpg, under_hood = TRUE)
x
wplay(notes(x$df$x_pitch, dur = 1 / 4, vol = 0.25), normalize = FALSE)

# specifying a different scale: e.g. equally-spaced arpeggio
new_scale <- c(16, 20, 23, 25) + rep(0:4 * 12, each = 4)
y <- scale_num(mtcars$mpg, new_scale = new_scale)
wplay(notes(y, dur = 1 / 4, vol = .25), normalize = FALSE)

}
