% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_parser.R
\name{scaling_parser}
\alias{scaling_parser}
\title{Interpret scaling string (helper function / Internal use)}
\usage{
scaling_parser(
  x,
  scaling = NULL,
  msg_varname = "x",
  non_neg = TRUE,
  warn_msgs = TRUE
)
}
\arguments{
\item{x}{numeric vector to be scaled}

\item{scaling}{string: description of desired scaling of data to pitch}

\item{msg_varname}{string: name of variable \code{x} to be used in messages}

\item{non_neg}{logical: if \code{TRUE}, scaled version of \code{x} (function output) is transformed to be non-negative}

\item{warn_msgs}{logical: if \code{TRUE} (default), prints messages for each scaling step}
}
\value{
list containing original \code{x}, scaled \code{x}, and other indicators/diagnostics
}
\description{
Helper function used in \code{sonify_scatter()}.  Uses \code{stringr} functions to parse \code{scaling} string.
}
\details{
uses \code{str_detect()} to parse \code{scaling} strings.  Looks for (in order):
\enumerate{
\item \code{bin} OR \code{quant}: place in equally-spaced bins OR bin by quantile.
\itemize{
\item If "bin" or "quant" detected, looks for integer as # of bins.
\item Integer must be after "bin" or "quant" but before "exp" (if "exp" specified); defaults to 5.
}
\item \code{rev}: reverse magnitude while keeping same range
\item \code{exp}: use variable as exponent
\itemize{
\item if "exp" detected, looks for integer/decimal to be the base (defaults to 2)
}
}

Mostly makes sense to use this function for volume and duration variables
(enforces non-negativity, reports 0's and NA's)
}
\examples{
\dontrun{
scaling_parser(x = airquality$Ozone, scaling = "binned 6 exp 1.5")
}
}
