% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify_box.R
\name{sonify_box}
\alias{sonify_box}
\title{Sonified Boxplot}
\usage{
sonify_box(
  data,
  var,
  group_var = NULL,
  tonic = 40,
  duration = NULL,
  phase_speed = 0.5,
  volume = 0.75,
  fs = 44100,
  progbar = TRUE
)
}
\arguments{
\item{data}{a dataframe}

\item{var}{: The variable to summarize with the sonified boxplot}

\item{group_var}{: If comparing distributions: the variable that splits the dataset into groups.}

\item{tonic}{int: the pkey with which the root the bottom note is mapped to.}

\item{duration}{int: the length of playback in seconds}

\item{phase_speed}{: numeric: the number of phases per second for each group.}

\item{volume}{numeric between 0 and 1: the relative volume of playback}

\item{fs}{int: The sampling rate for playback}

\item{progbar}{logical: if \code{TRUE} progress bar enabled}
}
\value{
2xn_samples matrix for playback.
}
\description{
A sound will slide pitchwise along the distribution of the data.
You can compare distributions by comparing the range of pitches being traveled.
}
\examples{
\dontrun{
dat <- data.frame(x = rnorm(100))
wplay(sonify_box(dat, x))
}
}
