% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify_hist.R
\name{sonify_hist}
\alias{sonify_hist}
\title{Sonified Histogram}
\usage{
sonify_hist(
  data,
  var,
  breaks = "Sturges",
  tonic = 40,
  inst = "sine",
  duration = 2,
  volume = 0.75,
  fs = 44100,
  progbar = TRUE
)
}
\arguments{
\item{data}{a dataframe}

\item{var}{the variable to create the sonified histogram for.}

\item{breaks}{int: the number of breaks. Sounds best if this entry is 2^n for some n.}

\item{tonic}{int: the pkey with which the root the bottom note is mapped to.}

\item{inst}{string: the name of the instrument used for playback}

\item{duration}{int: the length of playback in seconds}

\item{volume}{numeric between 0 and 1: the relative volume of playback}

\item{fs}{int: The sampling rate for playback}

\item{progbar}{logical: if \code{TRUE} progress bar enabled}
}
\value{
numeric vector representing the waveform of the sonified histogram
}
\description{
Map the relative frequencies to the pitches of the songs that are played in ascending order of your variable.
}
\examples{
\dontshow{wplayback <- FALSE}
dat <- data.frame(x = rnorm(100))
wplay(sonify_hist(dat, x))

}
