% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify_mod.R
\name{sonify_mod}
\alias{sonify_mod}
\title{Wrapper function for sonify_qq() and sonify_studres()}
\usage{
sonify_mod(
  lm_obj,
  mode = "qq",
  tonic_pkey = 34,
  alpha = 0.05,
  show_plot = FALSE
)
}
\arguments{
\item{lm_obj}{lm object returned by base R's \code{lm} function}

\item{mode}{string: specifies which sonification mode to use. Options: {"qq", "studres"}. "qq" produces an auditory QQ plot, and "studres" sonifies the externally studentized residuals.}

\item{tonic_pkey}{integer: the piano key of the tonic note used in the sonification. Both sonifications rely on a major scale.}

\item{alpha}{numeric: the false alarm rate (i.e., the p-value) to be used to detect observations that deviate from normality. Note that this will be a two-sided test, and \code{alpha} represents the cumulative false alarm rate.}

\item{show_plot}{boolean: whether or not a rudimentary plot of the sonification should also be shown. Default: FALSE.}
}
\value{
A 1 x n (if \code{mode == "studres"}) or 2 x n (if \code{mode == "qq"}) matrix representing the sound to be played. Generally, the user should pass this directly to \code{wplay()}.
}
\description{
Used to call \code{sonify_qq()} and \code{sonify_studres()} via a single function. Both of these subsidiary functions take an \code{lm} object as their input and produce a sonfied version of a standard model diagnostic.
}
\examples{
\dontrun{
# Run a linear regression on the mtcars data set
mod <- lm(mpg ~ 1 + cyl + hp + wt, data = mtcars)

# Play the QQ plot for this model fit
wplay(sonify_mod(mod, mode = "qq"))

# Play the studentized residuals for this model
wplay(sonify_mod(mod, mode = "studres"))
}
}
