% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify_scatter.R
\name{sonify_scatter}
\alias{sonify_scatter}
\title{Audio scatterplot}
\usage{
sonify_scatter(
  data,
  x_var,
  y_var,
  tonic_x = 40,
  tonic_y = 52,
  n_octaves_x = 2,
  n_octaves_y = 2,
  x_inst = "sine",
  y_inst = "piano",
  order_var = NULL,
  factor_var = NULL,
  volume = NULL,
  vol_scaling = NULL,
  duration = NULL,
  dur_scaling = NULL,
  speed = NULL,
  fs = 44100,
  verbose = TRUE,
  stereo = FALSE,
  under_hood = FALSE
)
}
\arguments{
\item{data}{data frame or tibble}

\item{x_var}{string: name of variable in supplied \code{data} to be used as the independent variable}

\item{y_var}{string: name of variable in supplied \code{data} to be used as the dependent variable}

\item{tonic_x}{numeric: piano key number of pitch beginning the scale that \code{x_var} is be mapped to}

\item{tonic_y}{numeric: piano key number of pitch beginning the scale that \code{y_var} is be mapped to}

\item{n_octaves_x}{numeric: number of octaves in scale used fo \code{x_var}}

\item{n_octaves_y}{numeric: number of octaves in scale used fo \code{y_var}}

\item{x_inst}{string: instrument used to play \code{x_var}}

\item{y_inst}{string: instrument used to play \code{y_var}}

\item{order_var}{string: name of variable used to order playback of notes (playback occurs in ascending order of \code{order_var}).  Default: \code{NULL}.  See details.}

\item{factor_var}{string (in quotes necessary): name of variable in supplied data to be used as factor variable.  Different levels in \code{factor_var} mapped to different instruments used in sonic representation of \code{y_var} (overrides \code{y_inst})}

\item{volume}{numeric, numeric vector, or name of variable in supplied data: volume of played sounds}

\item{vol_scaling}{string: scaling to be applied to \code{volume} variable.  See \code{\link{scaling_parser}}}

\item{duration}{numeric, numeric vector, or name of variable in supplied data: duration of played sounds}

\item{dur_scaling}{string: scaling to be applied to \code{duration} variable.  See \code{\link{scaling_parser}}}

\item{speed}{numeric: speed of playback.  See details.}

\item{fs}{integer: sampling rate (default: \code{44100}.  Sampling rate is dependent on user audio system, typically 44100 or 48000).}

\item{verbose}{logical: if \code{TRUE}, generates messages indicating variable mappings and transformations}

\item{stereo}{logical: if \code{TRUE}, generates stereo output, with \code{x_var} on left ear and \code{y_var} on right.}

\item{under_hood}{logical: if \code{TRUE}, returns list containing mappings used to construct output (wave vector/matrix is first element)}
}
\value{
numeric vector: playable audio scatterplot
}
\description{
Construct an audio scatterplot.
}
\details{
\itemize{
\item \code{order_var}: can specify a different variable in supplied data, "byrow", or, if unspecified (left \code{NULL}), uses \code{x_var} as \code{order_var}.
\item variable names for (\code{x_var}, \code{y_var}, \code{order_var}, \code{factor_var}, \code{volume}, \code{duration} need to be entered as quoted strings.
\item if \code{speed} left unspecified, defaults to \code{4} (so that \code{duration = 1} plays for 1/4 second instead of 1 second).  However, if \code{speed} unspecified but \code{dur_scaling} specified, then \code{speed} is set to 1.
}
}
\examples{
\dontrun{
wplay(
  sonify_scatter(
    data = mtcars,
    x_var = "mpg",
    y_var = "disp",
    factor_var = "gear"
  )
)


# airquality dataset; 
# order_var = "byrow" since dataset is ordered by day + month
wplay_controls <- TRUE
wplay(
  sonify_scatter(
    data = airquality,
    x_var = "Ozone",
    y_var = "Temp",
    order_var = "byrow",
    factor_var = "Month",
    duration = "Wind",
    dur_scaling = "binned 4 exp 2 rev"
  )
)
}
}
