% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify_studres.R
\name{sonify_studres}
\alias{sonify_studres}
\title{Sonification of externally studentized residuals.}
\usage{
sonify_studres(lm_obj, tonic_pkey = 34, alpha = 0.05, show_plot = FALSE)
}
\arguments{
\item{lm_obj}{lm object returned by base R's \code{lm} function}

\item{tonic_pkey}{integer: the piano key of the tonic note used in the sonification. Both sonifications rely on a major scale.}

\item{alpha}{numeric: the false alarm rate (i.e., the p-value) to be used to detect observations that deviate from normality. Note that this will be a two-sided test, and \code{alpha} represents the cumulative false alarm rate.}

\item{show_plot}{boolean: whether or not a rudimentary plot of the sonification should also be shown. Default: FALSE.}
}
\value{
A 1 x n matrix representing the sound to be played. Each chord is an observation. Observations that are below the alpha threshold will sound harmonious and will move up and down in pitch based on their relation to the mean. Observations that exceed the threshold will sound cacophonous.
}
\description{
short summary
}
\examples{
\dontshow{wplayback <- FALSE}
# Run a linear regression on the mtcars data set
mod <- lm(mpg ~ 1 + cyl + hp + wt, data = mtcars)

# Play the studentized residuals for this model fit
wplay(sonify_studres(mod))

}
