% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_norm.R
\name{wave_norm}
\alias{wave_norm}
\title{Normalize waves to ensure input to wplay() is valid}
\usage{
wave_norm(y)
}
\arguments{
\item{y}{Either a numeric vector or a 2xn_samples matrix}
}
\value{
a normalized wave of the same dimension as the input
}
\description{
Normalize waves to ensure input to wplay() is valid
}
\examples{
\dontrun{
wave_norm(sin(2 * pi * 60 * (1:1000) + 2 * pi * 100 * (1:1000)))
}
}
