% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplay.R
\name{wplay}
\alias{wplay}
\alias{cplay}
\alias{wplay_head}
\alias{wplay_tail}
\title{Playback an audio vector}
\usage{
wplay(
  y,
  vol = 0.25,
  normalize = TRUE,
  fs = 44100,
  file_path = NULL,
  file_type = "wav",
  wplayback,
  wplay_controls
)

cplay(y, vol = 0.25, normalize = TRUE, fs = 44100, n_bad = 5)

wplay_head(
  y,
  vol = 0.25,
  n_sec = 3,
  normalize = TRUE,
  fs = 44100,
  file_path = NULL,
  file_type = "wav"
)

wplay_tail(
  y,
  vol = 0.25,
  n_sec = 3,
  normalize = TRUE,
  fs = 44100,
  file_path = NULL,
  file_type = "wav"
)
}
\arguments{
\item{y}{either a n_samples length vector or a 2xn_samples matrix}

\item{vol}{numeric in [0, 1]: playback volume (default 0.25)}

\item{normalize}{logical: if \code{TRUE} (default) audio is normalized before playback.}

\item{fs}{integer: playback sampling rate (user system-dependent; typically 44100 or 48000)}

\item{file_path}{string: location/filename to save audio file when knitting to an html document}

\item{file_type}{string: file type of local copy when knitting html document (wav/mp3)}

\item{wplayback}{logical: if \code{TRUE} the audio is played.}

\item{wplay_controls}{logical: if \code{TRUE} the audio is played using in-console controls.  If \code{FALSE} places system in sleep during playback.}

\item{n_bad}{integer: number of invalid responses to console prompts before stopping.}

\item{n_sec}{numeric: number of seconds to play from beginning or end}
}
\value{
If called from script or in console, plays audio.  If knitting .Rmd document to html, saves sound file and inserts html audio tags into document
}
\description{
playback with system wait (default) or console controls.  When knitting .Rmd file to html, saves audio as .wav file and places html audio player in document.
}
\details{
\itemize{
\item \code{vol}: for these functions, \code{vol} is applied to the soundwave \emph{after} normalization (i.e. acts like a master volume for playback).
\item \code{file_path}: if unspecified, creates folder with same name as current .Rmd file and saves audio file using [timestamp].[\code{file_type}] as the filename.
\item \code{wplayback} can be set in parent environment / in console, i.e. can toggle audio playback on/off by creating a new variable \code{wplayback <- TRUE} or \code{FALSE}.   Useful when sourcing file / running all lines before, et cetera.
\item \code{wplay_controls} can be set in parent environment / in console.
}
}
\section{Functions}{
\itemize{
\item \code{cplay}: Audio playback with console controls (type "p" to pause, "r" to resume, "s" to stop).

\item \code{wplay_head}: Plays the first n_sec seconds of y

\item \code{wplay_tail}: Plays the last n_sec seconds of y
}}

\examples{
\dontrun{
x <- sin(1:100000/(3*pi))
wplay(x)

# two notes
y <- sin(1:100000/(5*pi))
wplay(x + y)

# no playback
wplay(x + y, wplayback = FALSE)
wplayback = FALSE
wplay(x)

# use parent environment to toggle console playback controls 
wplayback = TRUE
wplay_controls = TRUE
z <- sin(1:441000/(3*pi))  # playback takes 10 seconds.  Use console controls to stop.
wplay(z)
}
}
