% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wurr.R
\name{wurr}
\alias{wurr}
\title{Create continuous sliding sound.}
\usage{
wurr(
  hz = 440,
  phase_size = 1,
  phase_speed = 0.5,
  dur = 6,
  volume = 1,
  fs = 44100
)
}
\arguments{
\item{hz}{int: The middle frequency of the sound}

\item{phase_size}{int: The range of values being traversed}

\item{phase_speed}{: numeric: the number of phases per second for each group.}

\item{dur}{int: The duration of the sound in seconds}

\item{volume}{numeric between 0 and 1: the relative volume of playback}

\item{fs}{int: The sampling rate for playback}
}
\value{
a n_samples length sound vector
}
\description{
Create continuous sliding sound.
}
